Revert "refactor: JTAG関連の名前空間を修正"
This reverts commit 3469af89d6a7632666a9ab0368e20e6f50f08db1.
