+++ src/main/scala/board/jtag/type/VirtualInstruction.scala
@@ -1,4 +1,4 @@
-package board.types
+package board.jtag.types
 
 import chisel3._
 import support.types.DebugAccessDataKind
