+++ src/main/scala/board/jtag/type/VirtualJtagIO.scala
@@ -1,4 +1,4 @@
-package board.types
+package board.jtag.types
 
 import chisel3._
 import chisel3.internal.firrtl.Width
